Extracted CLI methods to separate file
